Organize private API in a different way  - Use the new API to update all behavior groups of an event type at    once
