+++ package.json
@@ -13,7 +13,7 @@
     "@redhat-cloud-services/frontend-components-notifications": "3.1.0",
     "@redhat-cloud-services/frontend-components-translations": "3.1.1",
     "@redhat-cloud-services/frontend-components-utilities": "3.1.2",
-    "@redhat-cloud-services/insights-common-typescript": "4.6.3",
+    "@redhat-cloud-services/insights-common-typescript": "4.7.0",
     "@redhat-cloud-services/rbac-client": "1.0.97",
     "axios": "0.21.1",
     "camelcase": "5.3.1",
@@ -26,7 +26,7 @@
     "immer": "9.0.2",
     "in-browser-download": "2.0.0",
     "lodash": "4.17.21",
-    "openapi2typescript": "0.1.3",
+    "openapi2typescript": "0.1.4",
     "p-limit": "3.0.2",
     "react": "17.0.2",
     "react-content-loader": "5.1.4",
@@ -92,7 +92,7 @@
     "eslint-plugin-simple-import-sort": "6.0.1",
     "eslint-plugin-testing-library": "3.10.0",
     "fetch-mock": "9.10.7",
-    "insights-common-typescript-dev": "0.1.13",
+    "insights-common-typescript-dev": "0.1.14",
     "jest": "26.6.3",
     "jest-mock": "26.1.0",
     "jest-transform-stub": "2.0.0",
@@ -103,7 +103,7 @@
     "karma-webpack": "5.0.0",
     "lock-dependencies": "0.0.2",
     "npm-run-all": "4.1.5",
-    "openapi2typescript-cli": "0.4.1",
+    "openapi2typescript-cli": "0.4.2",
     "postcss": "8.2.10",
     "prop-types": "15.7.2",
     "react-test-renderer": "17.0.2",
@@ -144,11 +144,13 @@
     "nightly": "npm run deploy",
     "proxy": "SPANDX_CONFIG=\"profiles/local-frontend-and-api.ts\" bash $PROXY_PATH/scripts/run.sh",
     "proxy-ui": "SPANDX_CONFIG=\"profiles/local-frontend.ts\" bash $PROXY_PATH/scripts/run.sh",
-    "schema": "npm-run-all schema:notifications schema:integrations",
+    "schema": "npm-run-all schema:notifications schema:integrations schema:private",
     "schema:notifications": "npm-run-all 'schema:notifications:generate -i http://localhost:8085/api/notifications/v1.0/openapi.json'",
     "schema:notifications:generate": "openapi2typescript -o src/generated/OpenapiNotifications.ts -ed -a react-fetching-library --explicit-types --no-strict",
     "schema:integrations": "npm-run-all 'schema:integrations:generate -i http://localhost:8085/api/integrations/v1.0/openapi.json'",
-    "schema:integrations:generate": "openapi2typescript -o src/generated/OpenapiIntegrations.ts -ed -a react-fetching-library --explicit-types --no-strict"
+    "schema:integrations:generate": "openapi2typescript -o src/generated/OpenapiIntegrations.ts -ed -a react-fetching-library --explicit-types --no-strict",
+    "schema:private": "npm-run-all 'schema:private:generate -i http://localhost:8085/api/private/v1.0/openapi.json'",
+    "schema:private:generate": "openapi2typescript -o src/generated/OpenapiPrivate.ts -ed -a react-fetching-library --explicit-types --no-strict"
   },
   "insights": {
     "appname": "notifications"
