+++ src/pages/Notifications/List/__tests__/Page.test.tsx
@@ -17,8 +17,8 @@ import {
 } from '../../../../../test/AppWrapper';
 import { waitForAsyncEvents } from '../../../../../test/TestUtils';
 import { VerboseErrorBoundary } from '../../../../../test/VerboseErrorBoundary';
-import { Schemas as SchemasBehaviorGroup } from '../../../../generated/OpenapiBehaviorGroups';
 import { Schemas } from '../../../../generated/OpenapiIntegrations';
+import { Schemas as SchemasBehaviorGroup } from '../../../../generated/OpenapiPrivate';
 import { linkTo } from '../../../../Routes';
 import { NotificationsListPage } from '../Page';
 import BehaviorGroup = SchemasBehaviorGroup.BehaviorGroup;
